chore(utils): use `Extract` generic for `ast-utils`' `predicates`' helper functions
