+++ packages/ast-spec/src/type/TSIntrinsicKeyword/spec.ts
